chore: rename type HandlerResult and move definition to response.ts file
