+++ src/libs/shared/src/types/form-state.ts
@@ -1,9 +1,7 @@
-import { BadResponse, GoodResponse } from "./response";
+import { HandlerResult } from "./response";
 
 export type FormState<T, P = unknown> = FieldError<T> | HandlerResult<P>;
 
 export type FieldError<T> = {
   fieldError: { [key in keyof T]: string[] };
 };
-
-export type HandlerResult<P> = { formResult: BadResponse | GoodResponse<P> }; // For the route handler specifically
