+++ src/libs/shared/src/types/response.ts
@@ -4,3 +4,5 @@ export type GoodResponse<T = NonNullable<unknown>> = {
 };
 
 export type BadResponse = { success: false; errors: string[] };
+
+export type HandlerResult<P> = { handlerResult: BadResponse | GoodResponse<P> }; // For the route handler specifically
